requirements.txt
alembic==0.9.1
appdirs==1.4.3
asn1crypto==0.24.0
azure-common==1.1.12
azure-nspkg==2.0.0
azure-storage-blob==1.3.0
azure-storage-common==1.3.0
azure-storage-nspkg==3.0.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.2.2
enum34==1.1.6
Flask==0.12.1
Flask-Migrate==2.0.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
futures==3.2.0
idna==2.7
ipaddress==1.0.22
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.6
MarkupSafe==1.0
packaging==16.8
psycopg2==2.7.1
pycparser==2.18
pyparsing==2.2.0
python-dateutil==2.7.3
python-editor==1.0.3
requests==2.19.1
six==1.10.0
SQLAlchemy==1.1.9
urllib3==1.23
waitress==1.0.1
Werkzeug==0.12.1